Se creó la clase abstracta Pokedex y se abtrae getPokemonsByNames() de PokedexOnline y PokedexOffline
