+++ server/cli.js
@@ -5,14 +5,14 @@ var pokemonComparator = new PokemonComparator();
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++
 // Para ejecutar de manera Online
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++
-// var Pokedex = require('./domain/pokedex')
-// var pokedex = new Pokedex();
+var Pokedex = require('./domain/pokedexOnline')
+var pokedex = new Pokedex();
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++
 // Para ejecutar de manera Offline
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-var PokedexOffline = require('./domain/pokedexOffline');
-var pokedex = new PokedexOffline();
+// var PokedexOffline = require('./domain/pokedexOffline');
+// var pokedex = new PokedexOffline();
 
 var CLI = function() {
 };
