+++ testAEjecutar.sh
@@ -1,4 +1,4 @@
 cd server/domain
 mocha pokemon.test.js
 mocha pokemonComparator.test.js
-mocha pokedex.test.js
+mocha pokedexOnline.test.js
